% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_dat.R
\docType{data}
\name{pop_dat}
\alias{pop_dat}
\title{Population dataset}
\format{
A tibble with X rows and Y variables:
\describe{
  \item{varname}{short desc}
}
}
\source{
\url{https://github.com/USDAForestService/FIESTA}
}
\usage{
pop_dat
}
\description{
Population dataset
}
\keyword{datasets}
