% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_dat.R
\docType{data}
\name{unit_dat}
\alias{unit_dat}
\title{Unit level dataset}
\format{
A tibble with X rows and Y variables:
\describe{
  \item{varname}{short desc}
}
}
\source{
\url{https://github.com/USDAForestService/FIESTA}
}
\usage{
unit_dat
}
\description{
Unit level dataset
}
\keyword{datasets}
